Move shader cache validator into its own file
